Add Cargo.lock to repository
By including the cargo.lock file we can ensure that all users
end up using the exact versions of dependencies as we did
during testing. Running `cargo update` will increase
version numbers to the latest released versions that are
compatible in the semantic versioning scheme, as defined
in the Cargo.toml file.